// Qanat.Web/src/app/shared/generated/model/usage-location-bulk-update-usage-location-type-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class UsageLocationBulkUpdateUsageLocationTypeDto { 
    UsageLocationTypeID: number;
    UsageLocationIDs: Array<number>;
    Note: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface UsageLocationBulkUpdateUsageLocationTypeDtoForm { 
    UsageLocationTypeID: FormControl<number>;
    UsageLocationIDs: FormControl<Array<number>>;
    Note: FormControl<string>;
}

export class UsageLocationBulkUpdateUsageLocationTypeDtoFormControls { 
    public static UsageLocationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static UsageLocationIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static Note = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
            ],
        }
    );
}